+++ firmware/tests/node/driver/MCP79510/test_driverMCP79510.c
@@ -1,7 +1,7 @@
 /**
  * @file   test_driverMCP79510.c
  * @Author Andreas Dahlberg (andreas.dahlberg90@gmail.com)
- * @date   2020-10-21 (Last edit)
+ * @date   2021-01-16 (Last edit)
  * @brief  Test suite for the MCP79510 driver.
  */
 /*
@@ -560,7 +560,7 @@ uint8_t __wrap_BCDToDecimal(uint8_t bcd_data)
     return data;
 }
 
-uint16_t __wrap_DecimalToBCD(uint8_t decimal_data)
+uint8_t __wrap_DecimalToBCD(uint8_t decimal_data)
 {
     uint16_t data = 0;
     uint8_t shift = 0;
@@ -573,5 +573,5 @@ uint16_t __wrap_DecimalToBCD(uint8_t decimal_data)
     }
     while (decimal_data > 0);
 
-    return data;
+    return (uint8_t)data;
 }
